Struct opengles_graphics::GlGraphics

pub struct GlGraphics { /* fields omitted */ }

Contains OpenGL data.

Methods

impl<'a> GlGraphics

fn new(opengl: OpenGL) -> Self

Creates a new OpenGL back-end.

Panics

If the OpenGL function pointers have not been loaded yet. See https://github.com/PistonDevelopers/opengl_graphics/issues/103 for more info.

fn viewport(&mut self, x: i32, y: i32, w: i32, h: i32)

Sets viewport with normalized coordinates and center as origin.

fn use_program(&mut self, program: GLuint)

Sets the current program only if the program is not in use.

fn clear_program(&mut self)

Unset the current program.

This forces the current program to be set on next drawing call.

fn use_draw_state(&mut self, draw_state: &DrawState)

Sets the current draw state, by detecting changes.

fn clear_draw_state(&mut self)

Unsets the current draw state.

This forces the current draw state to be set on next drawing call.

fn draw<F, U>(&mut self, viewport: Viewport, f: F) -> U where F: FnOnce(Context, &mut Self) -> U

Draws graphics.

fn has_texture_alpha(&self, _texture: &Texture) -> bool

Assume all textures has alpha channel for now.